SNOW-65947: Updated the create and drop schema script for internal Jenkins tests

+++ scripts/drop_schema.py
@@ -6,16 +6,22 @@ import os
 import sys
 import snowflake.connector
 
-travis_job_id = os.getenv('TRAVIS_JOB_ID')
-appveyor_job_id = os.getenv('APPVEYOR_BUILD_ID')
-if not travis_job_id and not appveyor_job_id:
-    print("[WARN] The environment variable TRAVIS_JOB_ID or APPVEYOR_BUILD_ID is not set. No test schema will be created.")
-    sys.exit(1)
+test_schema = None
 
-if travis_job_id:
-  test_schema = 'TRAVIS_JOB_{0}'.format(travis_job_id)
-else:
-  test_schema = 'APPVEYOR_BUILD_{0}'.format(appveyor_job_id)
+if 'TRAVIS_JOB_ID' in os.environ:
+    job_id = os.getenv('TRAVIS_JOB_ID')
+    test_schema = 'TRAVIS_JOB_{0}'.format(job_id)
+
+if 'APPVEYOR_BUILD_ID' in os.environ:
+    job_id = os.getenv('APPVEYOR_BUILD_ID')
+    test_schema = 'APPVEYOR_BUILD_{0}'.format(job_id)
+
+if 'BUILD_TAG' in os.environ:
+    test_schema = os.getenv('SNOWFLAKE_TEST_SCHEMA')
+
+if test_schema is None:
+    print("[WARN] The environment variable TRAVIS_JOB_ID or APPVEYOR_BUILD_ID or JENKINS_BUILD_TAG is not set. No test schema will be created.")
+    sys.exit(1)
 
 params = {
     'account': os.getenv("SNOWFLAKE_TEST_ACCOUNT"),
@@ -34,7 +40,8 @@ protocol=os.getenv("SNOWFLAKE_TEST_PROTOCOL")
 if protocol:
     params['protocol'] = protocol
 
-con = snowflake.connector.connect(**params)
-con.cursor().execute("drop schema if exists {0}".format(test_schema))
+if not test_schema.lower() in ['testschema', 'public']:
+    con = snowflake.connector.connect(**params)
+    con.cursor().execute("drop schema if exists {0}".format(test_schema))
 
 sys.exit(0)
